+++ meta_form_validators/form_validators/blood_results_form_validators.py
@@ -1,7 +1,8 @@
 from django import forms
 from edc_blood_results import BloodResultsFormValidatorMixin
-from edc_constants.constants import FASTING
+from edc_constants.constants import FASTING, YES
 from edc_form_validators.form_validator import FormValidator
+from edc_glucose.utils import validate_glucose_as_millimoles_per_liter
 from edc_lab_panel.panels import (
     blood_glucose_panel,
     blood_glucose_poc_panel,
@@ -20,18 +21,29 @@ class BaseBloodResultsGluFormValidator:
     def reportables_evaluator_options(self):
         if not self.cleaned_data.get("fasting"):
             raise forms.ValidationError({"fasting": "This field is required."})
-        fasting = True if self.cleaned_data.get("fasting") == FASTING else False
+        fasting = (
+            True
+            if (
+                (self.cleaned_data.get("fasting") == FASTING)
+                or (self.cleaned_data.get("fasting") == YES)
+            )
+            else False
+        )
         return dict(fasting=fasting)
+
+    def evaluate_value(self, field_name):
+        if field_name == "glucose_value":
+            validate_glucose_as_millimoles_per_liter("glucose", self.cleaned_data)
 
 
 class BloodResultsGluFormValidator(
-    BloodResultsFormValidatorMixin, BaseBloodResultsGluFormValidator, FormValidator
+    BaseBloodResultsGluFormValidator, BloodResultsFormValidatorMixin, FormValidator
 ):
     panel = blood_glucose_panel
 
 
 class BloodResultsGluPocFormValidator(
-    BloodResultsFormValidatorMixin, BaseBloodResultsGluFormValidator, FormValidator
+    BaseBloodResultsGluFormValidator, BloodResultsFormValidatorMixin, FormValidator
 ):
     panel = blood_glucose_poc_panel