+++ meta_subject/admin/blood_results/blood_results_hba1c_admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 from edc_blood_results.admin import BloodResultsModelAdminMixin
-from edc_blood_results.fieldsets import BloodResultPanel
+from edc_blood_results.fieldsets import BloodResultFieldset
 from edc_lab_panel.panels import hba1c_panel
 
 from ...admin_site import meta_subject_admin
@@ -14,7 +14,7 @@ from ..modeladmin import CrfModelAdmin
 @admin.register(BloodResultsHba1c, site=meta_subject_admin)
 class BloodResultsHba1cAdmin(BloodResultsModelAdminMixin, CrfModelAdmin):
     form = BloodResultsHba1cForm
-    fieldsets = BloodResultPanel(hba1c_panel).fieldsets
+    fieldsets = BloodResultFieldset(hba1c_panel).fieldsets
 
 
 # from django.contrib import admin