+++ meta_subject/admin/blood_results/blood_results_lft_admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 from edc_blood_results.admin import BloodResultsModelAdminMixin
-from edc_blood_results.fieldsets import BloodResultPanel
+from edc_blood_results.fieldsets import BloodResultFieldset
 from edc_lab_panel.panels import lft_panel
 
 from ...admin_site import meta_subject_admin
@@ -12,7 +12,7 @@ from ..modeladmin import CrfModelAdmin
 @admin.register(BloodResultsLft, site=meta_subject_admin)
 class BloodResultsLftAdmin(BloodResultsModelAdminMixin, CrfModelAdmin):
     form = BloodResultsLftForm
-    fieldsets = BloodResultPanel(lft_panel, model_cls=BloodResultsLft).fieldsets
+    fieldsets = BloodResultFieldset(lft_panel, model_cls=BloodResultsLft).fieldsets
 
 
 # from django.contrib import admin