+++ meta_subject/admin/blood_results/blood_results_rft_admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 from edc_blood_results.admin import BloodResultsModelAdminMixin
-from edc_blood_results.fieldsets import BloodResultPanel
+from edc_blood_results.fieldsets import BloodResultFieldset
 from edc_lab_panel.panels import rft_panel
 
 from ...admin_site import meta_subject_admin
@@ -12,7 +12,7 @@ from ..modeladmin import CrfModelAdmin
 @admin.register(BloodResultsRft, site=meta_subject_admin)
 class BloodResultsRftAdmin(BloodResultsModelAdminMixin, CrfModelAdmin):
     form = BloodResultsRftForm
-    fieldsets = BloodResultPanel(rft_panel, model_cls=BloodResultsRft).fieldsets
+    fieldsets = BloodResultFieldset(rft_panel, model_cls=BloodResultsRft).fieldsets
 
 
 # from django.contrib import admin