+++ meta_subject/forms/glucose_form.py
@@ -1,28 +1,19 @@
 from django import forms
 from edc_crf.modelform_mixins import CrfModelFormMixin
 from edc_form_validators.form_validator import FormValidator
-from meta_screening.form_validators import GlucoseFormValidatorMixin
-from meta_screening.forms import validate_glucose_as_millimoles_per_liter
+from edc_glucose.form_validators import GlucoseFormValidatorMixin
 
 from ..models import Glucose
 
 
 class GlucoseFormValidator(GlucoseFormValidatorMixin, FormValidator):
     def clean(self):
-        self.validate_ifg()
-        self.validate_ogtt()
-        self.validate_ogtt_dates()
-        self.validate_glucose_dates()
+        self.validate_glucose_testing_matrix()
 
 
 class GlucoseForm(CrfModelFormMixin, forms.ModelForm):
     form_validator_cls = GlucoseFormValidator
 
-    def clean(self):
-        cleaned_data = super().clean()
-        validate_glucose_as_millimoles_per_liter(cleaned_data)
-        return cleaned_data
-
     class Meta:
-        model = Glucose
+        model = Glucose  # "Glucose (IFG, OGTT)"
         fields = "__all__"